+++ pkg/providers/botman/data_akamai_botman_bot_analytics_settings_values_test.go
@@ -1,0 +1,53 @@
+package botman
+
+import (
+	"testing"
+
+	"github.com/akamai/terraform-provider-akamai/v10/internal/edgegrid"
+	"github.com/akamai/terraform-provider-akamai/v10/pkg/common/testutils"
+	"github.com/hashicorp/terraform-plugin-testing/helper/resource"
+)
+
+func TestDataBotAnalyticsSettingsValues(t *testing.T) {
+	t.Parallel()
+	t.Run("DataBotAnalyticsSettingsValues", func(t *testing.T) {
+		t.Parallel()
+		client := edgegrid.NewTestClient()
+		mockGetConfigVersion(client.APPSEC)
+
+		response := map[string]interface{}{
+			"values": []interface{}{
+				map[string]interface{}{"testKey": "testValue1"},
+				map[string]interface{}{"testKey": "testValue2"},
+				map[string]interface{}{"testKey": "testValue3"},
+				map[string]interface{}{"testKey": "testValue4"},
+				map[string]interface{}{"testKey": "testValue5"},
+			},
+		}
+		expectedJSON := `
+{
+	"values": [
+		{"testKey":"testValue1"},
+		{"testKey":"testValue2"},
+		{"testKey":"testValue3"},
+		{"testKey":"testValue4"},
+		{"testKey":"testValue5"}
+	]
+}`
+		client.BotMan.On("GetBotAnalyticsSettingsValues",
+			testutils.MockContext,
+		).Return(response, nil)
+		resource.UnitTest(t, resource.TestCase{
+			ProtoV6ProviderFactories: testutils.NewTestProtoV6ProviderFactory(client, NewSubprovider()),
+			Steps: []resource.TestStep{
+				{
+					Config: testutils.LoadFixtureString(t, "testdata/TestDataBotAnalyticsSettingsValues/basic.tf"),
+					Check: resource.ComposeAggregateTestCheckFunc(
+						resource.TestCheckResourceAttr("data.akamai_botman_bot_analytics_settings_values.test", "json", compactJSON(expectedJSON))),
+				},
+			},
+		})
+
+		client.BotMan.AssertExpectations(t)
+	})
+}